Prevent /dev/ipmi0 file descriptor leak when using DCMI to gather power readings.
This removes the thread local approach in favour of creating a new DCMI context
each time a power reading is made. Previously IPMI contexts could remain
allocated after threads ended, resulting in a memory and file descriptor leak.
The file descriptor leak was particularly problematic as a threshold is
reached, the kernel driver no longer allows further files to be opened,
resulting in both a failure of Slurm to be able to read power readings, as well
as all other usage of `/dev/ipmi0` by other processes.