Further work to mitigate counter leak causing failures within power accounting gathering.
Additional diagnostics showed that the counter is decremented even when it has
not been incremented, resulting in it containing a negative value. For an
unsigned integer, this is interpreted as a very positive value and leads to
failure of the conditional test for being less than 10 -- and then decremented
even further.

This patch prevents that from happening. It also increases the thread limit to
64 (one quarter of the thread pool) which should allow up to 64 jobs to be
started at around the same time before limits are reached; this number is still
somewhat arbitrary, but should suit the majority of use cases except in
extremely high core count systems which are processing large numbers of low
core jobs.